42Projects: CommonCore

+++ 09 Minishell(beforeMerge)/srcs/execute/execute.c
@@ -6,11 +6,11 @@
 /*   By: min-kang <minguk.gaang@gmail.com>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/16 15:54:12 by min-kang          #+#    #+#             */
-/*   Updated: 2022/03/15 18:41:55 by min-kang         ###   ########.fr       */
+/*   Updated: 2022/03/20 15:17:07 by min-kang         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../includes/minishell.h"
+#include "minishell.h"
 
 int	cmd_execute(t_node *node, int fd_in, int fd_out, char **envp)
 {
@@ -26,31 +26,6 @@ int	cmd_execute(t_node *node, int fd_in, int fd_out, char **envp)
 	execve(cmd_path, node->left->args, envp);
 	write(fd_out, "", 1);
 	exit (127);
-}
-
-int	proc_child(t_node *node, char **envp, int fd_in, int *fd)
-{
-	close(fd[0]);
-	if (node->right && node->right->node_type > 1)
-		cmd_execute(node->left, fd_in, fd[1], envp);
-	else
-		cmd_execute(node, fd_in, 1, envp);
-	return (1);
-}
-
-int	proc_parent(t_node *node, char **envp, int *fd, pid_t pid)
-{
-	int	res;
-	int	res_child;
-
-	close(fd[1]);
-	res = execute_loop(node->right, envp, fd[0]);
-	close(fd[0]);
-	waitpid(pid, &res_child, 0);
-	res_child = WEXITSTATUS(res_child);
-	if (node->right && node->right->node_type > 1)
-		return (res);
-	return (res_child);
 }
 
 int	execute_loop(t_node *node, char **envp, int fd_in)
@@ -78,8 +53,20 @@ int	parse_execute(t_token *begin, int index, char **envp)
 	int		res;
 
 	node = parser(begin, index);
+	if (!node->root)
+	{
+		free(node);
+		return (0);
+	}
+	else if (!node->root->left)
+	{
+		free_node(node->root);
+		free(node);
+		return (0);
+	}
 	res = execute_loop(node->root, envp, 0);
-	unlink(HEREDOC);
 	free_node(node->root);
+	free(node);
+	unlink(HEREDOC);
 	return (res);
 }
